detail::pow template, msvc fixes

+++ tests/tests.cpp
@@ -318,30 +318,30 @@ TEST(LiteralsPrefixes) {
     static_assert("rm / m"_u  == 1e-27);
     static_assert("qm / m"_u  == 1e-30);
 
-    static_assert("m / Qm"_u  == 1 / 1e30 );
-    static_assert("m / Rm"_u  == 1 / 1e27 );
-    static_assert("m / Ym"_u  == 1 / 1e24 );
-    static_assert("m / Zm"_u  == 1 / 1e21 );
-    static_assert("m / Em"_u  == 1 / 1e18 );
-    static_assert("m / Pm"_u  == 1 / 1e15 );
-    static_assert("m / Tm"_u  == 1 / 1e12 );
-    static_assert("m / Gm"_u  == 1 / 1e9  );
-    static_assert("m / Mm"_u  == 1 / 1e6  );
-    static_assert("m / km"_u  == 1 / 1e3  );
-    static_assert("m / hm"_u  == 1 / 1e2  );
-    static_assert("m / dam"_u == 1 / 1e1  );
-    static_assert("m / dm"_u  == 1 / 1e-1 );
-    static_assert("m / cm"_u  == 1 / 1e-2 );
-    static_assert("m / mm"_u  == 1 / 1e-3 );
-    static_assert("m / um"_u  == 1 / 1e-6 );
-    static_assert("m / nm"_u  == 1 / 1e-9 );
-    static_assert("m / pm"_u  == 1 / 1e-12);
-    static_assert("m / fm"_u  == 1 / 1e-15);
-    static_assert("m / am"_u  == 1 / 1e-18);
-    static_assert("m / zm"_u  == 1 / 1e-21);
-    static_assert("m / ym"_u  == 1 / 1e-24);
-    static_assert("m / rm"_u  == 1 / 1e-27);
-    static_assert("m / qm"_u  == 1 / 1e-30);
+    static_assert("m / Qm"_u  == double(1 / 1e30L ));
+    static_assert("m / Rm"_u  == double(1 / 1e27L ));
+    static_assert("m / Ym"_u  == double(1 / 1e24L ));
+    static_assert("m / Zm"_u  == double(1 / 1e21L ));
+    static_assert("m / Em"_u  == double(1 / 1e18L ));
+    static_assert("m / Pm"_u  == double(1 / 1e15L ));
+    static_assert("m / Tm"_u  == double(1 / 1e12L ));
+    static_assert("m / Gm"_u  == double(1 / 1e9L  ));
+    static_assert("m / Mm"_u  == double(1 / 1e6L  ));
+    static_assert("m / km"_u  == double(1 / 1e3L  ));
+    static_assert("m / hm"_u  == double(1 / 1e2L  ));
+    static_assert("m / dam"_u == double(1 / 1e1L  ));
+    static_assert("m / dm"_u  == double(1 / 1e-1L ));
+    static_assert("m / cm"_u  == double(1 / 1e-2L ));
+    static_assert("m / mm"_u  == double(1 / 1e-3L ));
+    static_assert("m / um"_u  == double(1 / 1e-6L ));
+    static_assert("m / nm"_u  == double(1 / 1e-9L ));
+    static_assert("m / pm"_u  == double(1 / 1e-12L));
+    static_assert("m / fm"_u  == double(1 / 1e-15L));
+    static_assert("m / am"_u  == double(1 / 1e-18L));
+    static_assert("m / zm"_u  == double(1 / 1e-21L));
+    static_assert("m / ym"_u  == double(1 / 1e-24L));
+    static_assert("m / rm"_u  == double(1 / 1e-27L));
+    static_assert("m / qm"_u  == double(1 / 1e-30L));
 }
 
 TEST(LiteralsTime) {
@@ -355,8 +355,8 @@ TEST(Literals3) {
     static_assert(1e6 * "mm"_u == "km"_u);
     static_assert(1e6 * "Pa"_u == "MPa"_u);
 
-    static_assert("mL"_u / "cm3"_u == (1e-3 * 1e-3 / (1e-2 * 1e-2 * 1e-2)));
-    static_assert("L"_u / "dm3"_u == (1e-3 / (0.1 * 0.1 * 0.1)));
+    static_assert("mL"_u / "cm3"_u == (double(1e-3L * 1e-3L) / double(1e-2L * 1e-2L * 1e-2L)));
+    static_assert("L"_u / "dm3"_u == (double(1e-3L) / double(0.1L * 0.1L * 0.1L)));
 
     static_assert("Pa"_u == "N/m2"_u);
 
@@ -370,7 +370,7 @@ TEST(Literals3) {
 
     static_assert("ha"_u == 1e4 * "m2"_u);
     static_assert("ha"_u == "hm2"_u);
-    static_assert("ha"_u / "yd2"_u == (1e4 / (0.9144 * 0.9144)));
+    static_assert("ha"_u / "yd2"_u == (1e4 / double(0.9144L * 0.9144L)));
 
     static_assert("t"_u == 1e3 * "kg"_u);
